requirements.txt
psycopg2-binary
pandas
pyarrow
sqlalchemy
PyMySQL
minio